Hax packing a punch!
Added script in the package.json file so you are able to package the application for your respective machine. For macs run package-mac, for PCs run package-win, for linux run package-linux.

small note: when you are on a PC make sure you are running the command prompt with admin provilages. I encountered an error (no icon found) when I was not in admin